requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file requirements.txt requirements.in
#
aiohttp==1.0.5            # via discord.py
async-timeout==1.2.1      # via aiohttp
chardet==3.0.4            # via aiohttp
crontab==0.21.3
discord.py==0.16.8
multidict==3.1.3          # via aiohttp
websockets==3.3           # via discord.py
